Consolidate main menu controllers
