--- MonoChrome/Assets/Scripts/Core/CoreMainMenuController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c244145d414ad2444b62b14d1b503458
