+++ MonoChrome/Assets/Scripts/Core/SystemMigrator.cs
@@ -175,14 +175,14 @@ namespace MonoChrome.Core
                 }
             }
             
-            // CoreMainMenuController 생성 (MainMenu 씬에서만)
+            // MainMenuController 생성 (MainMenu 씬에서만)
             if (UnityEngine.SceneManagement.SceneManager.GetActiveScene().name == "MainMenu")
             {
-                if (FindObjectOfType<CoreMainMenuController>() == null)
+                if (FindObjectOfType<MonoChrome.MainMenuController>() == null)
                 {
-                    GameObject coreMainMenuObj = new GameObject("CoreMainMenuController");
-                    coreMainMenuObj.AddComponent<CoreMainMenuController>();
-                    LogMessage("  ✓ CoreMainMenuController 생성됨");
+                    GameObject coreMainMenuObj = new GameObject("MainMenuController");
+                    coreMainMenuObj.AddComponent<MonoChrome.MainMenuController>();
+                    LogMessage("  ✓ MainMenuController 생성됨");
                 }
             }
             
@@ -292,7 +292,7 @@ namespace MonoChrome.Core
             LogMessage("Core 매니저들:");
             LogMessage($"  CoreGameManager: {(FindObjectOfType<CoreGameManager>() != null ? "있음" : "없음")}");
             LogMessage($"  CoreUIManager: {(FindObjectOfType<CoreUIManager>() != null ? "있음" : "없음")}");
-            LogMessage($"  CoreMainMenuController: {(FindObjectOfType<CoreMainMenuController>() != null ? "있음" : "없음")}");
+            LogMessage($"  MainMenuController: {(FindObjectOfType<MonoChrome.MainMenuController>() != null ? "있음" : "없음")}");
             
             // UI 구조 확인
             Canvas canvas = FindObjectOfType<Canvas>();
